backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chromadb==1.1.0
click==8.2.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.7
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
estimator==0.0.10
executing==2.2.0
fastapi==0.104.1
fastjsonschema==2.21.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.55.6
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.9.2
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
langsmith==0.4.29
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml_dtypes==0.5.1
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.0.8
narwhals==2.3.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.1.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.106.1
opencv-python==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.15.0
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
platformdirs==4.3.6
polars==1.35.2
polars-runtime-32==1.35.2
posthog==5.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==5.29.4
psutil==6.1.1
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.13.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.27.0
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.227
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.24.0
vaderSentiment==3.3.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
google-generativeai==0.8.3
